refactoring of parser and formatter
